LayoutTests/fast/mediastream/device-change-event-2-expected.txt

CONSOLE MESSAGE: A MediaStreamTrack ended due to a capture failure
CONSOLE MESSAGE: A MediaStreamTrack ended due to a capture failure

PASS 'devicechange' event fired when device list changes
PASS 'devicechange' events fired quickly are coalesced
PASS 'devicechange' event is not fired when the document is not visible
